[FRCV-33] Refactor type imports and relocate type files (#32)
## [FRCV-33] Description
Moved type definition files from 'src/types' to more context-specific directories within 'src/containers', 'src/models', and 'src/services/ClusterManager'. Updated all relevant imports to reflect the new file locations, improving code organization and maintainability.
