+++ src/containers/types/ai-service/assistant-message.types.ts
