+++ src/models/EventEndpoint.ts
@@ -1,6 +1,6 @@
 import IORedis from 'ioredis';
 import { AI, Cluster, Core, InstanceBase, Microservice, Thread } from '../services';
-import { EndpointSetup } from '../types/EventEndpoint.types';
+import { EndpointSetup } from './types/EventEndpoint.types';
 
 const REDIS_URL = process.env.REDIS_URL || 'redis://localhost:6379';
 const ioRedis = new IORedis(REDIS_URL);
