+++ src/models/types/EventEndpoint.types.ts
